feat: add subscription (#9)
Adding event.Subscription that will handle envelope data for events. It
constrains for all events to have the same data/schema and validates the
name of the event matches the one expected. Invalid data is discarded
automatically. To handle more flexible scenarios use RawSubscription.

Since we don't want to focus on allowing different events on the same
subscription there is no support for it for now. If it becomes necessary
my idea would be to build a different Subscription that supports muxing
like the http.ServeMux but multiplexing event names instead of URL
paths.