chore: add request_id to each event (#30)
To handle situations where clients may send multiple requests with the
same trace_id. We have some more details about this on our handbook, the
idea is to have something that is GUARANTEED to be unique and per
request/event/msg, independent of client behavior/etc, so filtering logs
per request context is always trivial (and trace IDs can be used to
correlate things across service boundaries).

Events are not request, but we want to have a common/unique name...and
we went with request_id for now.