feat: add event.NewMessage function (#34)
If you want to test a MessageHandler you need to be able to create a
Message to pass as a parameter. The current design doesn't allow it. Im
not sure about the design of Message...my idea was to push into the
direction of making the message immutable (quite a good property when
debugging odd scenarios on message handling), but doing that in Go is
not trivial (requires copying the []byte when accessing it..or else it
is mutable anyway...)...so Im not sure the abstraction is worth its
cost...Im open to alternatives..but keeping in line with the current
abstraction would be to allow for it to be built with a function (since
the fields are private).