feat: add xhttp client retrier (#41)
Introduce the xhttp package. x may stand for:

* eXperimental
* eXtended
* EXTREME !!!
* XGuiga (the best option of course)

Take your pick :-). The idea is to avoid name clashing with stdlib
packages, but extend them with things that are useful for us and more
specific.

The main design idea of the retrier is to be a drop-in replacement for
the stdlib http.Client (but for now we only implement Do, since it is
what we use most of the time AFAIK).

Basic usage is:

```go
	client := xhttp.NewRetrierClient(&http.Client{})
        client.Do(req)
```

The same proxying can be done later for metrics/caching/etc.