feat:tracing: add fine grained control over request stats (#70)
Some services handle way too much requests (dozens of millions per day).
Logging each single request on the
INFO level is starting to generate considerable costs (and we rarely
need those logs..if ever). In order to be
able to change this behavior Im introducing a new type and a callback
for the request stats. The current default
behavior will be maintained.