feat: add Subscription.ReceiveN (#73)
This will make batching easier. We usually deal with one event at a time
in a "handler" fashion, but sometimes we need to batch N events together
in memory (usually to optimize some processing, which is the case right
now). This PR introduces the basics for using subscriptions with a
Receive method, similar to the one on Go CDK but with our event envelope
and also a ReceiveN that will help with batching.