feat: add xerrgroup (#80)
Very basic extension to errgroup. Have been using conc for this, but
realized that what I want is very simple. Simpler than what they have
there, the API has a lot of different types/constructors/error
modes/cancellation/etc...I never remember which one I want, because what
I want is pretty much errgroup but I want auto collection of results
because Im lazy and I do this a lot.

It is a design choice to not allow zero values to be valid, I like being
able to use the zero value but sadly I would not be able to support
WithContext since it returns a pointer and we cant copy the underlying
errgroup. If we drop WithContext it would be possible to make it work
with zero values, but WithContext can be important in scenarios where
you want to be able to cancel all subtasks if one of them failed.