feat: add xhttp.NewJSONRequest (#86)
My idea is to add a layer as small as possible just to make it easier to
handle JSON on HTTP request/responses. Usually errors are disjointed
from responses (different fields), so I'm thinking on always creating a
type that is the union of whatever you expect as response + what you
would expect if it is an error. After checking the status code then you
can validate each part of the parsed data, the error for an error status
code and the valid response for a success. This allows for more
sophisticated things to be built on top, like something that already
checks the status code and hides HTTP entirely, but not sure that is
necessary/desirable once we have something that already makes it
considerably less annoying to receive JSON (no more reading/error
handling/parsing for every request).

The current design interoperates with anything that creates an
[http.Request] and the [xhttp.Client] interface is implemented by Go's
[http.Client] and our retrier (basically, everything just works
:man_shrugging: ).