redesign: xjson.DynGet with value,err instead of 3 return vals (#94)
Sometimes I do the 3 return values things for some stupid reason because
I almost always regret it later. In a lot of use cases I don't care if
it failed for an error or was just not found (and for debugging/logging
the error message will be enough to make it clear), handling the result
gets considerably more annoying (I might be traumatized by io.EOF...but
not found issues are not the same thing I think).