feat: add tracing context request ID (#97)
We already had the request_id but only for logs. The request ID is good
to identify specific request/responses when the same trace_id may be
used (like in a retry). It is a more limited version of a span ID for
tracing, but it is what we have today (we may have something more
complete someday). With this we can add the request ID as metadata in an
event related to a request and then easily correlate them (the trace ID
is not enough...but maybe that is debatable...but right now I also don't
have a whole lot of time).