feat(xhttp): retry on 504 Gateway Timeout error. (#105)
I was on the fence if makes sense to implement a specific test for this
or we assume it works the same for all retryable status code.
If you think I should test this, let me know.

I know we have the `RetrierWithStatuses()` option but this case seems
too be a perfect case as retrying could resolve the issue and it's a
real case happening at Birdie that would require update several clients
to use the option if not added here.

Signed-off-by: i4k <t.nateldemoura@gmail.com>